backend/requirements.txt
fastapi>=0.95.0,<1.0.0
uvicorn[standard]>=0.20.0,<0.25.0
sqlalchemy[asyncio]>=2.0.0,<2.1.0
asyncpg>=0.27.0,<0.29.0
alembic>=1.10.0,<1.12.0
pydantic[email]>=1.10.7,<2.0.0
python-jose[cryptography]>=3.3.0,<3.4.0
passlib[bcrypt]>=1.7.4,<1.8.0
python-multipart>=0.0.6,<0.0.7
websockets>=11.0.0,<12.0.0
boto3>=1.26.0,<1.29.0
requests>=2.28.0,<2.32.0 # Often useful, might be implicitly needed
jinja2>=3.1.0,<3.2.0
authlib>=1.2.0,<1.3.0 # For OAuth
httpx>=0.24.0,<0.26.0 # Required by Authlib
# GStreamer is a system dependency, not a pip package usually.
# Ensure GStreamer 1.x development libraries are installed on the system.
# For rate limiting and other core features, no specific external package seems used beyond standard libraries. 